refactor: reorganize collab-adapter directory
trying to get it ready for packaging
