+++ client/src/lib/collab-adapter/index.ts
@@ -3,7 +3,12 @@ import type { Awareness } from 'y-protocols/awareness'
 
 // Main adapter factory
 export { default as createFirebaseYWebrtcAdapter } from './adapter'
-export type { AdapterOptions, AdapterHandle } from './adapter'
+export type { 
+  AdapterOptions, 
+  AdapterHandle,
+  AdapterEvents,
+  SignalData
+} from './core/types'
 
 // Configuration utilities
 export { 
@@ -22,6 +27,31 @@ export type {
   ConnectionState 
 } from './utils/config'
 
+// Constants
+export {
+  DEFAULT_MAX_DIRECT_PEERS,
+  PEER_PRESENCE_TIMEOUT_MS,
+  PEER_ID_DISPLAY_LENGTH,
+  MAX_MESSAGE_BUFFER_SIZE,
+  MESSAGE_BUFFER_RETENTION_MS,
+  IDLE_PEER_TIMEOUT_MS,
+  MAX_MEMORY_BUFFER_BYTES,
+  MAX_AWARENESS_STATES,
+  CLEANUP_INTERVAL_MS,
+  HEARTBEAT_INTERVAL_MS,
+  STALE_CONNECTION_TIMEOUT_MS,
+  MEMORY_CHECK_INTERVAL_MS,
+  MIN_VISIBILITY_UPDATE_INTERVAL,
+  DEFAULT_SYNC_INTERVAL_MS,
+  AWARENESS_THROTTLE_MS,
+  COMPRESSION_THRESHOLD,
+  USE_NATIVE_COMPRESSION,
+  STUN_SERVERS,
+} from './utils/constants'
+
+// Compression utilities
+export { compressData, decompressData } from './utils/compression'
+
 // Persistence utilities (for advanced users)
 export { 
   loadDocumentFromFirebase,
@@ -35,20 +65,12 @@ export type { DocumentSnapshot } from './firebase/persistence'
 export { announcePresence, stopAnnouncingPresence } from './firebase/presence'
 export type { PeerInfo } from './firebase/presence'
 
-// Re-export types from config and cluster
+// Core peer manager (for advanced users)
+export { SimplePeerManager } from './core/peer-manager'
+
+// Re-export types from config and cluster for convenience
 import type { ConnectionState } from './utils/config'
 import type { PeerInfo } from './firebase/presence'
-
-// Event types for the adapter
-export type AdapterEvents = {
-  'connection-state-changed': { state: ConnectionState }
-  'peer-joined': { peerId: string; user: PeerInfo }
-  'peer-left': { peerId: string }
-  'document-persisted': { docId: string; version: number }
-  'error': { error: Error; context: string }
-  'sync-completed': { docId: string; updateSize: number }
-  'awareness-updated': { peerId: string; user: PeerInfo }
-}
 
 // Enhanced adapter interface types
 export type YDocumentAdapter = {
@@ -66,12 +88,13 @@ export type YDocumentAdapter = {
     awarenessStates: number
   }
   forceGarbageCollection(): void
-  on<K extends keyof AdapterEvents>(event: K, callback: (data: AdapterEvents[K]) => void): void
-  off<K extends keyof AdapterEvents>(event: K, callback: (data: AdapterEvents[K]) => void): void
+  forcePersist(): Promise<void>
+  on<K extends keyof import('./core/types').AdapterEvents>(event: K, callback: (data: import('./core/types').AdapterEvents[K]) => void): void
+  off<K extends keyof import('./core/types').AdapterEvents>(event: K, callback: (data: import('./core/types').AdapterEvents[K]) => void): void
 }
 
 // Re-export AdapterOptions for factory
-import type { AdapterOptions } from './adapter'
+import type { AdapterOptions } from './core/types'
 
 export type AdapterFactory<TOptions = AdapterOptions> = {
   create(options: TOptions): Promise<YDocumentAdapter>
